fixed-requirements.txt
# Fixed requirements.txt

# Keep crewai at version 0.130.0 as specified
crewai==0.130.0 
crewai-tools==0.47.1

# Core FastAPI and web framework dependencies
fastapi==0.110.3
uvicorn[standard]>=0.15.0
python-multipart

# AI/ML and LLM dependencies
openai==1.30.5
langchain>=0.1.0
langchain-openai
langchain-community
langchain-core==0.1.52

# PDF processing dependencies (FIXED: Added missing PDF processing libraries)
PyPDF2>=3.0.0
pypdf>=3.0.0
pdfplumber>=0.7.0

# Google Cloud and AI dependencies  
google-ai-generativelanguage==0.6.4
google-api-core==2.10.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0

# Data processing and utilities
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0

# Environment and configuration
python-dotenv>=0.19.0
click==8.1.7

# Template and JSON processing
Jinja2==3.1.4
jsonschema==4.22.0

# Monitoring and telemetry
langsmith==0.1.67
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0

# Runtime dependencies
onnxruntime==1.18.0
oauthlib==3.2.2
protobuf==4.25.3
pydantic==1.10.13
pydantic_core==2.8.0

# Development and testing (Optional - uncomment if needed)
# pytest>=6.0.0
# pytest-asyncio
# black
# flake8
# mypy